+++ spec/associations/device_auth/subscriptions_spec.rb
@@ -11,21 +11,21 @@ describe Yt::Associations::Subscriptions, :device_app do
 
       context 'that I am not subscribed to' do
         before { channel.unsubscribe }
-        it { expect(channel.subscribed?).to be_false }
-        it { expect(channel.subscribe!).to be_true }
+        it { expect(channel.subscribed?).to be false }
+        it { expect(channel.subscribe!).to be_truthy }
       end
 
       context 'that I am subscribed to' do
         before { channel.subscribe }
-        it { expect(channel.subscribed?).to be_true }
-        it { expect(channel.unsubscribe!).to be_true }
+        it { expect(channel.subscribed?).to be true }
+        it { expect(channel.unsubscribe!).to be_truthy }
       end
     end
 
     # NOTE: This test is just a reflection of YouTube irrational behavior of
     # raising a 500 error when you try to subscribe to your own channel, rather
     # than a more logical 4xx error. Hopefully this will get fixed and this
-    # test removed.
+    # code (and test) removed.
     context 'given my own channel' do
       let(:channel) { Yt::Channel.new id: $account.channel.id, auth: $account }