+++ spec/models/rating_spec.rb
@@ -5,9 +5,8 @@ describe Yt::Rating do
   subject(:rating) { Yt::Rating.new }
 
   describe '#update' do
-    before { rating.stub(:do_update).and_yield }
+    before { expect(rating).to receive(:do_update).and_yield }
 
-    it { expect(rating.update :like).to be_true }
     it { expect{rating.update :like}.to change{rating.rating} }
   end
 end