+++ HISTORY.md
@@ -14,6 +14,7 @@ v0.5 - 2014/05/16
 * New Earning collection to retrieve estimated earning for YouTube-partnered channels
 * Rename error classes so they match the corresponding Net::HTTP errors (e.g. Unauthorized)
 * Separate Error class for 403 Error
+* Retry once YouTube earning queries that return error 400 "Invalid query. Query did not conform to the expectations"
 
 v0.4 - 2014/05/09
 --------------------